Fix richiesti in PR
