+++ netaddiction_account/__init__.py
@@ -2,3 +2,4 @@
 # License LGPL-3.0 or later (https://www.gnu.org/licenses/lgpl).
 
 from . import models
+from . import wizard
