+++ netaddiction_account/wizard/__init__.py
@@ -1,0 +1,1 @@
+from . import check_csv_cod
